+++ src/i18n.js
@@ -1,7 +1,7 @@
 module.exports = {
     vi: {
         title: "TXA AUTO ACCEPT",
-        version: "v7.4.2",
+        version: "v7.4.3",
         monitor: "GIÁM SÁT",
         config: "CẤU HÌNH",
         shield: "LÁ CHẮN",
@@ -81,7 +81,7 @@ module.exports = {
     },
     en: {
         title: "TXA AUTO ACCEPT",
-        version: "v7.4.2",
+        version: "v7.4.3",
         monitor: "MONITOR",
         config: "CONFIG",
         shield: "SHIELD",